refactor: use a new file to storage the api types
